requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.1